refactor: extrai as seções da landing /melhor-idade para componentes
O React Doctor apontou no-giant-component na PR #1316: o componente passou de
300 linhas. As demais landings do repo ficam entre 79 e 133 linhas, com as
seções em components/landings/<nome>/ — este commit alinha a /melhor-idade a
esse padrão, seguindo ConsultoriaLandingSections como modelo.

- pages/landings/MelhorIdadeLanding.tsx: 392 → 72 linhas (SEO, schemas, FAQ
  e composição das seções)
- components/landings/melhor-idade/MelhorIdadeLandingSections.tsx: Nav, Hero,
  Pillars, Rhythm, Advisory, Content e FinalCta como componentes exportados

Sem mudança de marcação, estilo ou copy — o e2e da credencial Pastore segue
passando nos 5 navegadores sem alteração.

O helper de animação virou buildRise (não useRise): não é um hook React e o
prefixo `use` induziria as rules-of-hooks ao erro.
